+++ x-core/test/x/core/client_on_response_test.rb
@@ -3,10 +3,11 @@ require_relative "../../test_helper"
 module X
   class ClientOnResponseTest < Minitest::Test
     cover Client
+    cover StreamingClient
 
     def setup
       @responses = []
-      @client = Client.new(max_stream_reconnects: 0, on_response: ->(response) { @responses << response })
+      @client = Client.new(on_response: ->(response) { @responses << response })
     end
 
     def test_on_response_receives_every_response
@@ -28,8 +29,8 @@ module X
       stub_request(:get, "https://api.twitter.com/2/tweets/sample/stream")
         .to_return(body: "{\"data\":{\"id\":\"1\"},\"includes\":{\"users\":[{\"id\":\"2\"}]}}\r\n\r\n{\"data\":{\"id\":\"3\"}}\r\n", headers: {"x-rate-limit-limit" => "50", "x-rate-limit-remaining" => "49", "x-rate-limit-reset" => "1"})
       events = []
-      client = Client.new(max_stream_reconnects: 0, on_response: ->(response) { events << [response.resource_count, response.rate_limit.remaining, response.uri.path] })
-      client.stream("tweets/sample/stream") { |post| events << post.dig("data", "id") }
+      client = Client.new(on_response: ->(response) { events << [response.resource_count, response.rate_limit.remaining, response.uri.path] })
+      client.streaming(max_reconnects: 0).stream("tweets/sample/stream") { |post| events << post.dig("data", "id") }
 
       assert_equal [[2, 49, "/2/tweets/sample/stream"], "1", [1, 49, "/2/tweets/sample/stream"], "3"], events
     end
@@ -37,7 +38,7 @@ module X
     def test_on_response_receives_a_failed_stream_before_the_error
       stub_request(:get, "https://api.twitter.com/2/tweets/search/stream").to_return(status: 429, body: '{"title":"Too Many Requests"}', headers: {"Content-Type" => "application/json"})
 
-      assert_raises(TooManyRequests) { @client.stream("tweets/search/stream") { flunk "unexpected yield" } }
+      assert_raises(TooManyRequests) { @client.streaming(max_reconnects: 0).stream("tweets/search/stream") { flunk "unexpected yield" } }
       assert_equal [[:get, 429, '{"title":"Too Many Requests"}']], @responses.map { |response| [response.http_method, response.status, response.body] }
     end